# backend/src/doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "PsicoGestion API",
    "description": "Documentación completa de endpoints con autenticación JWT."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT en formato: Bearer <token>"
    }
  },
...
  "paths": {
...
    "/api/citas/paciente/{paciente_id}": {
      "get": {
        "tags": ["Citas"],
        "summary": "Citas por paciente",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "paciente_id", "in": "path", "required": true, "type": "integer" }],
        "responses": { "200": { "description": "Citas", "schema": { "type": "array", "items": { "$ref": "#/definitions/Cita" } } } }
      }
    },
...
    "/api/expedientes/paciente/{paciente_id}": {
      "post": {
        "tags": ["Expedientes"],
        "summary": "Crear expediente",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "paciente_id", "in": "path", "required": true, "type": "integer" }, { "in": "body", "name": "body", "schema": { "$ref": "#/definitions/Expediente" } }],
        "responses": { "201": { "description": "Expediente creado" } }
      },
      "put": {
        "tags": ["Expedientes"],
        "summary": "Actualizar expediente",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "paciente_id", "in": "path", "required": true, "type": "integer" }, { "in": "body", "name": "body", "schema": { "$ref": "#/definitions/Expediente" } }],
        "responses": { "200": { "description": "Expediente actualizado" } }
      }
    },
...
    "/api/pacientes": {
      "get": {
        "tags": ["Pacientes"],
        "summary": "Listar pacientes",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Pacientes", "schema": { "type": "array", "items": { "$ref": "#/definitions/Paciente" } } } }
      },
      "post": {
        "tags": ["Pacientes"],
        "summary": "Crear paciente",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "in": "body", "name": "body", "schema": { "$ref": "#/definitions/Paciente" } }],
        "responses": { "201": { "description": "Paciente creado" } }
      }
    },
    "/api/pacientes/{id}": {
      "get": {
        "tags": ["Pacientes"],
        "summary": "Detalle de paciente",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }],
        "responses": { "200": { "description": "Paciente", "schema": { "$ref": "#/definitions/Paciente" } } }
      },
      "put": {
        "tags": ["Pacientes"],
        "summary": "Actualizar paciente",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }, { "in": "body", "name": "body", "schema": { "$ref": "#/definitions/Paciente" } }],
        "responses": { "200": { "description": "Paciente actualizado" } }
      },
      "delete": {
        "tags": ["Pacientes"],
        "summary": "Eliminar paciente (soft-delete)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }],
        "responses": { "200": { "description": "Paciente inactivado" } }
      }
    },
    "/api/pacientes/activos": {
      "get": {
        "tags": ["Pacientes"],
        "summary": "Pacientes activos",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Pacientes", "schema": { "type": "array", "items": { "$ref": "#/definitions/Paciente" } } } }
      }
    },
    "/api/pacientes/sin-asignar": {
      "get": {
        "tags": ["Pacientes"],
        "summary": "Pacientes sin asignar",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Pacientes", "schema": { "type": "array", "items": { "$ref": "#/definitions/Paciente" } } } }
      }
    },
    "/api/pacientes/candidatos-alta": {
      "get": {
        "tags": ["Pacientes"],
        "summary": "Pacientes candidatos a alta",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Pacientes", "schema": { "type": "array", "items": { "$ref": "#/definitions/Paciente" } } } }
      }
    },
    "/api/pacientes/{id}/marcar-no-aprobado": {
      "post": {
        "tags": ["Pacientes"],
        "summary": "Marcar paciente como no aprobado",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }, { "in": "body", "name": "body", "schema": { "type": "object", "additionalProperties": true } }],
        "responses": { "200": { "description": "Paciente actualizado" } }
      }
    },
    "/api/pacientes/{id}/no-aprobar-alta": {
      "post": {
        "tags": ["Pacientes"],
        "summary": "No aprobar alta de paciente",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }, { "in": "body", "name": "body", "schema": { "type": "object", "additionalProperties": true } }],
        "responses": { "200": { "description": "Alta no aprobada" } }
      }
    },
...
    "/api/sesiones/paciente/{paciente_id}": {
      "get": {
        "tags": ["Sesiones"],
        "summary": "Sesiones por paciente",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "paciente_id", "in": "path", "required": true, "type": "integer" }],
        "responses": { "200": { "description": "Sesiones", "schema": { "type": "array", "items": { "$ref": "#/definitions/Sesion" } } } }
      }
    },
...
    "/api/estadisticas/paciente/{paciente_id}": {
      "get": {
        "tags": ["Estadisticas"],
        "summary": "Estadísticas de paciente",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "paciente_id", "in": "path", "required": true, "type": "integer" }],
        "responses": { "200": { "description": "Estadísticas", "schema": { "$ref": "#/definitions/Estadisticas" } } }
      }
    },
...
  },
  "definitions": {
...
    "Paciente": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "nombre": { "type": "string" },
        "apellido": { "type": "string" },
        "email": { "type": "string" },
        "telefono": { "type": "string" },
        "fecha_nacimiento": { "type": "string" },
        "genero": { "type": "string" },
        "direccion": { "type": "string" },
        "estado": { "type": "string" },
        "activo": { "type": "boolean" },
        "notas": { "type": "string" },
        "fundacion_id": { "type": "integer" },
        "motivo_consulta": { "type": "string" }
      }
    },
...
    "Cita": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "paciente_id": { "type": "integer" },
        "psicologo_id": { "type": "integer" },
        "becario_id": { "type": "integer" },
        "fecha": { "type": "string" },
        "hora_inicio": { "type": "string" },
        "hora_fin": { "type": "string" },
        "estado": { "type": "string" },
        "tipo": { "type": "string" }
      }
    },
    "Sesion": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "paciente_id": { "type": "integer" },
        "psicologo_id": { "type": "integer" },
        "fecha": { "type": "string" },
        "notas": { "type": "string" }
      }
    },
    "Expediente": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "paciente_id": { "type": "integer" },
        "psicologo_id": { "type": "integer" },
        "motivo_consulta": { "type": "string" }
      }
    },
...
    "Estadisticas": {
      "type": "object",
      "additionalProperties": true
    },
...
  }
}